move cart discount operations to correct namespace
