+++ commercetools.NET/CartDiscounts/CartDiscountQueryResult.cs
@@ -2,7 +2,7 @@
 using commercetools.Common;
 using Newtonsoft.Json;
 
-namespace commercetools.CartDiscount
+namespace commercetools.CartDiscounts
 {
     public class CartDiscountQueryResult: PagedQueryResult
     {
