+++ commercetools.NET/CartDiscounts/CartDiscountTarget.cs
@@ -1,9 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
-namespace commercetools.CartDiscount
+namespace commercetools.CartDiscounts
 {
     public class CartDiscountTarget
     {
