+++ commercetools.NET/CartDiscounts/CartDiscountValue.cs
@@ -4,7 +4,7 @@ using commercetools.Common;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Converters;
 
-namespace commercetools.CartDiscount
+namespace commercetools.CartDiscounts
 {
     public class CartDiscountValue
     {
