+++ commercetools.NET/CartDiscounts/UpdateActions/ChangeIsActive.cs
@@ -1,7 +1,7 @@
 ﻿using commercetools.Common;
 using Newtonsoft.Json;
 
-namespace commercetools.CartDiscount.UpdateActions
+namespace commercetools.CartDiscounts.UpdateActions
 {
     public class ChangeIsActive: UpdateAction
     {
