+++ commercetools.NET/CartDiscounts/UpdateActions/ChangeName.cs
@@ -1,7 +1,7 @@
 ﻿using commercetools.Common;
 using Newtonsoft.Json;
 
-namespace commercetools.CartDiscount.UpdateActions
+namespace commercetools.CartDiscounts.UpdateActions
 {
     public class ChangeName: UpdateAction
     {
