+++ commercetools.NET/CartDiscounts/UpdateActions/SetDescription.cs
@@ -1,7 +1,7 @@
 ﻿using commercetools.Common;
 using Newtonsoft.Json;
 
-namespace commercetools.CartDiscount.UpdateActions
+namespace commercetools.CartDiscounts.UpdateActions
 {
     public class SetDescription: UpdateAction
     {
