+++ commercetools.NET/commercetools.NET.csproj
@@ -57,24 +57,24 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CartDiscount\CartDiscount.cs" />
-    <Compile Include="CartDiscount\CartDiscountDraft.cs" />
-    <Compile Include="CartDiscount\CartDiscountManager.cs" />
-    <Compile Include="CartDiscount\CartDiscountQueryResult.cs" />
-    <Compile Include="CartDiscount\CartDiscountTarget.cs" />
-    <Compile Include="CartDiscount\CartDiscountValue.cs" />
-    <Compile Include="CartDiscount\Enums.cs" />
-    <Compile Include="CartDiscount\Extensions.cs" />
-    <Compile Include="CartDiscount\UpdateActions\ChangeCartPredicate.cs" />
-    <Compile Include="CartDiscount\UpdateActions\ChangeIsActive.cs" />
-    <Compile Include="CartDiscount\UpdateActions\ChangeName.cs" />
-    <Compile Include="CartDiscount\UpdateActions\ChangeRequiresDiscountCode.cs" />
-    <Compile Include="CartDiscount\UpdateActions\ChangeSortOrder.cs" />
-    <Compile Include="CartDiscount\UpdateActions\ChangeTarget.cs" />
-    <Compile Include="CartDiscount\UpdateActions\ChangeValue.cs" />
-    <Compile Include="CartDiscount\UpdateActions\SetDescription.cs" />
-    <Compile Include="CartDiscount\UpdateActions\SetValidFrom.cs" />
-    <Compile Include="CartDiscount\UpdateActions\SetValidUntil.cs" />
+    <Compile Include="CartDiscounts\CartDiscount.cs" />
+    <Compile Include="CartDiscounts\CartDiscountDraft.cs" />
+    <Compile Include="CartDiscounts\CartDiscountManager.cs" />
+    <Compile Include="CartDiscounts\CartDiscountQueryResult.cs" />
+    <Compile Include="CartDiscounts\CartDiscountTarget.cs" />
+    <Compile Include="CartDiscounts\CartDiscountValue.cs" />
+    <Compile Include="CartDiscounts\Enums.cs" />
+    <Compile Include="CartDiscounts\Extensions.cs" />
+    <Compile Include="CartDiscounts\UpdateActions\ChangeCartPredicate.cs" />
+    <Compile Include="CartDiscounts\UpdateActions\ChangeIsActive.cs" />
+    <Compile Include="CartDiscounts\UpdateActions\ChangeName.cs" />
+    <Compile Include="CartDiscounts\UpdateActions\ChangeRequiresDiscountCode.cs" />
+    <Compile Include="CartDiscounts\UpdateActions\ChangeSortOrder.cs" />
+    <Compile Include="CartDiscounts\UpdateActions\ChangeTarget.cs" />
+    <Compile Include="CartDiscounts\UpdateActions\ChangeValue.cs" />
+    <Compile Include="CartDiscounts\UpdateActions\SetDescription.cs" />
+    <Compile Include="CartDiscounts\UpdateActions\SetValidFrom.cs" />
+    <Compile Include="CartDiscounts\UpdateActions\SetValidUntil.cs" />
     <Compile Include="Carts\ExternalLineItemTotalPrice.cs" />
     <Compile Include="Carts\UpdateActions\AddCustomLineItemAction.cs" />
     <Compile Include="Carts\UpdateActions\AddDiscountCodeAction.cs" />
